pyproject.toml
[project]
name = "puddle-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "mcp[cli]>=1.22.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
